requirements.txt
# requirements.txt
streamlit==1.30.0
google-generativeai==0.3.1
pandas==2.1.3
python-dotenv==1.0.0
